remove_old_lineage
