+++ dinky-web/package.json
@@ -56,7 +56,6 @@
     "@xterm/xterm": "^5.5.0",
     "antd": "5.21.0",
     "antd-style": "^3.7.1",
-    "butterfly-dag": "^4.3.29",
     "classnames": "^2.5.1",
     "dayjs": "^1.11.11",
     "echarts": "^5.5.0",
@@ -79,7 +78,6 @@
     "react-fast-marquee": "^1.6.5",
     "react-grid-layout": "^1.4.4",
     "react-helmet-async": "^2.0.5",
-    "react-lineage-dag": "^2.0.36",
     "react-lottie": "^1.2.4",
     "react-markdown": "^9.0.1",
     "react-resizable-panels": "^2.1.4",
