+++ dinky-web/src/pages/DevOps/JobDetail/JobLineage/index.tsx
@@ -17,17 +17,15 @@
  *
  */
 
-import LineageGraph from '@/components/LineageGraph';
 import { queryDataByParams } from '@/services/BusinessCrud';
 import { API_CONSTANTS } from '@/services/endpoints';
 import { LineageDetailInfo } from '@/types/DevOps/data';
 import { l } from '@/utils/intl';
-import { connect } from '@umijs/max';
 import { Card, Result } from 'antd';
 import React, { useEffect } from 'react';
-import 'react-lineage-dag/dist/index.css';
+import {Lineage} from "@/pages/DataStudio/Toolbar/Service/Lineage";
 
-const JobLineage: React.FC<connect> = (props) => {
+const JobLineage = (props:{jobDetail:{id:number}}) => {
   const {
     jobDetail: { id: jobInstanceId }
   } = props;
@@ -48,9 +46,9 @@ const JobLineage: React.FC<connect> = (props) => {
 
   return (
     <>
-      <Card hoverable bodyStyle={{ height: '100%' }} style={{ height: parent.innerHeight - 180 }}>
+      <Card hoverable styles={{body:{height:'100%'}} } style={{ height: parent.innerHeight - 180 }}>
         {lineageData && (lineageData.tables.length !== 0 || lineageData.relations.length !== 0) ? (
-          <LineageGraph lineageData={lineageData} refreshCallBack={queryLineageData} />
+          <Lineage data={lineageData} />
         ) : (
           <Result style={{ height: '100%' }} status='warning' title={l('lineage.getError')} />
         )}
